requirements.txt
autopep8==1.4.4
Click==7.0
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
numpy==1.17.4
opencv-python==4.1.2.30
opencv-utils==0.0.2
Pillow==6.2.1
pycodestyle==2.5.0
pygifsicle==1.0.1
